So, I finally found the problem with FMRadio!  It was a problem with the application, not the compiler (well, maybe, read on).  It was a problem with the input filter:
void->float filter FloatOneSource {

  int x;

  init {
    x = 0;
  }

  work push 1 pop 0 {
    push(x++);
  }
}

Anyone see the problem?  Come on.

The front-end should really check that the type of push statement is the
same type as the filter output type.  I added a check to the raw backend
and I will check it in.  I will also add a check to Kopi2SIR if I can.

This was a problem in FloatOneSource also.

+++ streams/apps/benchmarks/fm/streamit/FloatOneSource.java
@@ -16,11 +16,11 @@ import streamit.library.*;
 
 class FloatOneSource extends Filter
 {
-    int x;
+    float x;
     public void init ()
     {
         output = new Channel(Float.TYPE, 1);
-	x = 0;
+	x = 0.0f;
     }
     public void work()
     {
